+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - Renamed package to `com.tobiasdiez.easybind`.
 - Split `MonadicBinding` into a part which is really concerned with optionals (new `ObervableOptionalValue`) and one which provides helper methods for a fluent interface for normal bindings (new `EasyBinding`).
 - Renamed `flatMap` to `mapObservable` to avoid clashes with the corresponding method in `Optional`.
+- The `mapObservable` and `selectProperty` now invoke the mapper even if the source observable value holds a `null` value. The corresponding methods in `ObservableOptionalValue` allow for a more convenient `null` handling.
 - Renamed subscribe methods that accept listeners to `listen`, which invoke the given listener every time the value changes. In contrast, the `subscribe` method also invokes the given subscriber upon registration. 
 - Renamed `EasyBind.listBind` to `bindContent` in order to align with the naming of `Bindings.bindContent`.